# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  title: string;
  category: string;
  amount: number;
  date: string;
  /** @nullable */
  note: string | null;
  createdAt?: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface TransactionInput {
  type: TransactionInputType;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  category: string;
  /** @exclusiveMinimum 0 */
  amount: number;
  date: string;
  note?: string;
}

export type TransactionUpdateType = typeof TransactionUpdateType[keyof typeof TransactionUpdateType];


export const TransactionUpdateType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface TransactionUpdate {
  type?: TransactionUpdateType;
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  category?: string;
  /** @exclusiveMinimum 0 */
  amount?: number;
  date?: string;
  note?: string;
}

export interface CategoryTotal {
  category: string;
  amount: number;
  percentage: number;
}

export interface MonthlyTotal {
  month: string;
  income: number;
  expenses: number;
  profit: number;
}

export interface Dashboard {
  month: string;
  totalIncome: number;
  totalExpenses: number;
  netProfit: number;
  incomeChange: number;
  expenseChange: number;
  savingsRate: number;
  topExpenseCategory: string;
  expenseBreakdown: CategoryTotal[];
  monthlyTrend: MonthlyTotal[];
  recentTransactions: Transaction[];
}

export interface AssistantSummaryInput {
  /** @pattern ^[0-9]{4}-[0-9]{2}$ */
  month: string;
  question?: string;
}

export interface AssistantSummary {
  message: string;
  highlights: string[];
}

export type ListTransactionsParams = {
/**
 * @pattern ^[0-9]{4}-[0-9]{2}$
 */
month?: string;
type?: ListTransactionsType;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType];


export const ListTransactionsType = {
  income: 'income',
  expense: 'expense',
} as const;

export type GetDashboardParams = {
/**
 * @pattern ^[0-9]{4}-[0-9]{2}$
 */
month?: string;
};