# lib/api-zod/src/generated/types/assistantSummaryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantSummaryInput {
  /** @pattern ^[0-9]{4}-[0-9]{2}$ */
  month: string;
  question?: string;
}