# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryTotal } from './categoryTotal';
import type { MonthlyTotal } from './monthlyTotal';
import type { Transaction } from './transaction';

export interface Dashboard {
  month: string;
  totalIncome: number;
  totalExpenses: number;
  netProfit: number;
  incomeChange: number;
  expenseChange: number;
  savingsRate: number;
  topExpenseCategory: string;
  expenseBreakdown: CategoryTotal[];
  monthlyTrend: MonthlyTotal[];
  recentTransactions: Transaction[];
}