# lib/api-zod/src/generated/types/monthlyTotal.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyTotal {
  month: string;
  income: number;
  expenses: number;
  profit: number;
}