# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from './transactionType';

export interface Transaction {
  id: number;
  type: TransactionType;
  title: string;
  category: string;
  amount: number;
  date: Date;
  /** @nullable */
  note: string | null;
  createdAt?: Date;
}