# lib/api-zod/src/generated/types/transactionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionInputType } from './transactionInputType';

export interface TransactionInput {
  type: TransactionInputType;
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  category: string;
  /** @exclusiveMinimum 0 */
  amount: number;
  date: Date;
  note?: string;
}